requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
psycopg[binary]==3.2.3
twilio==9.2.3
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
